Refactor account balance as functional
